% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{get_mozhi_engines}
\alias{get_mozhi_engines}
\title{Get available engines from Mozhi frontend}
\usage{
get_mozhi_engines(instance = NULL)
}
\arguments{
\item{instance}{The Mozhi instance URL you want to use}
}
\description{
View and store the available engines from the Mozhi frontend you've chosen.
}
\examples{
\dontrun{
get_mozhi_engines()
}
}
