% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{gtranslate}
\alias{gtranslate}
\title{Translate strings using gtranslate frontend}
\usage{
gtranslate(from, to, str)
}
\arguments{
\item{from}{Source language. Specify it with the code following the ISO 639 standard}

\item{to}{Specify it with the code following the ISO 639 standard}

\item{str}{Text to translate}
}
\description{
gtranslate is a privacy-respecting frontend for the Google Translate engine.
}
\examples{
\dontrun{
gtranslate(from = "en", to = "es", str = "hello")
}
}
