% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{lingva}
\alias{lingva}
\title{Translate strings using Lingva frontend}
\usage{
lingva(from, to, str, instance = NULL)
}
\arguments{
\item{from}{Source language. Specify it with the code following the ISO 639 standard}

\item{to}{Specify it with the code following the ISO 639 standard}

\item{str}{Text to translate}

\item{instance}{The Lingva instance URL you want to use. If it's not set, it will ask you to select one}
}
\description{
Lingva is a privacy-respecting frontend for the Google Translate engine.
}
\examples{
\dontrun{
lingva(from = "en", to = "es", str = "hello")
}
}
