% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{mozhi}
\alias{mozhi}
\title{Translate strings using Mozhi frontend}
\usage{
mozhi(from, to, str, instance = NULL, engine = NULL)
}
\arguments{
\item{from}{Source language. Specify it with the code following the ISO 639 standard}

\item{to}{Specify it with the code following the ISO 639 standard}

\item{str}{Text to translate}

\item{instance}{The Mozhi instance URL you want to use}

\item{engine}{The translation engine you want to use. You can see which ones are
available using \code{get_mozhi_engines()} function}
}
\description{
Mozhi is a privacy-respecting frontend that was born as a fork of SimplyTranslate
but now has many more features and supports many more engines.
}
\examples{
\dontrun{
mozhi(from = "en", to = "es", str = "hello")
}
}
