% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{mozhi_all}
\alias{mozhi_all}
\title{Translate with all Mozhi available engines.}
\usage{
mozhi_all(from, to, str, instance = NULL)
}
\arguments{
\item{from}{Source language. Specify it with the code following the ISO 639 standard}

\item{to}{Specify it with the code following the ISO 639 standard}

\item{str}{Text to translate}

\item{instance}{The Mozhi instance URL you want to use}
}
\description{
This function returns a dataframe with the translated string for each
engine.
}
\examples{
\dontrun{
mozhi_all(from = "es", to = "en", str = text)
}
}
