% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{simplytranslate}
\alias{simplytranslate}
\title{Translate strings using SimplyTranslate}
\usage{
simplytranslate(from, to, str, instance = NULL, engine = NULL)
}
\arguments{
\item{from}{Source language. Specify it with the code following the ISO 639 standard}

\item{to}{Specify it with the code following the ISO 639 standard}

\item{str}{Text to translate}

\item{instance}{The SimplyTranslate URL you want to use}

\item{engine}{The translation engine you want to use: "google", "reverso" or "inciba"}
}
\description{
SimplyTranslate has support for three translation engines: Google Translate,
inCIBA and Reverso. You can specify which one do you want to use session-wide
or every time you execute the function. This function also helps with the selection
of the SimplyTranslate instance you want to use.
}
\examples{
\dontrun{
simplytranslate(from = "en", to = "es", str = "hello", instance = "https://simplytranslate.org)")
}
}
